+++ plugins/fused-ops-and-kernels/tests/test_fused_ops.py
@@ -1,275 +1,401 @@
 import torch
-from bitsandbytes.nn.modules import Linear4bit
-from auto_gptq.nn_modules.qlinear.qlinear_tritonv2 import QuantLinear
-from peft import get_peft_model, LoraConfig
-from peft.tuners.lora.bnb import Linear4bit as LoraBNBLinear4bit
-from peft.tuners.lora.gptq import QuantLinear as LoraGPTQLinear4bit
 from copy import deepcopy
 from transformers import AutoConfig
 from transformers.models.llama.modeling_llama import LlamaAttention
-from peft import LoraConfig, get_peft_model
+
+from peft import LoraConfig
+from peft.tuners.lora.bnb import Linear4bit as LoraBNBLinear4bit
+from peft.tuners.lora.gptq import QuantLinear as LoraGPTQLinear4bit
+
 from fms_acceleration_foak.models.model_patcher import patch_model, patch_model_summary
-from transformers import AutoConfig, AutoModelForCausalLM, BitsAndBytesConfig
-from types import MethodType
-from functools import partial
+
+from itertools import product
+
+import pytest
 
 BNB = "bitsandbytes"
 GPTQ = "auto_gptq"
 
+TEST_MODELS = {
+    BNB: (
+        "TinyLlama/TinyLlama-1.1B-Chat-v0.3",  
+        LlamaAttention,
+        ['q_proj', 'k_proj', 'v_proj', 'o_proj']
+    ),
+    GPTQ: (
+        "TheBloke/TinyLlama-1.1B-Chat-v0.3-GPTQ",
+        LlamaAttention,
+        ['q_proj', 'k_proj', 'v_proj', 'o_proj']
+    ),
+}
+
+LORA_QUANTIZED_CLASSES = {
+    BNB: LoraBNBLinear4bit,
+    GPTQ: LoraGPTQLinear4bit,
+}
+
+ADAPTER_NAME = 'default'
+
+LOSS_TOL = 1e-3
+ALLCLOSE_RTOL = 1e-3
+ALLCLOSE_ATOL = 1e-4
+
+FLOAT16 = 'float16'
+DTYPES = [FLOAT16]
+
+DROPOUTS = [0, 0.1, 0.5]
+
+# r, lora_alpha
+LORA_PARAMS = [(8, 1.0)]
+
+# bs, seqlen, hiddim
+SIZES = [(1, 128, 2048)]
+
 # set a fixed dropout to match outputs between runs
 class DummyDropout(torch.nn.Module):
-    def __init__(self, dropout_mask=None):
+
+    dropout_mask: torch.tensor = None
+
+    def __init__(self):
         super().__init__()
-        self.dropout_mask = dropout_mask
 
     def forward(self, X):
+
+        # may raise if dropout mask not set
         if self.training:
             return X * self.dropout_mask
         return X
 
-def wrap_peft(module, peft_cls, r, lora_alpha, lora_dropout):
-    module.q_proj = peft_cls(module.q_proj, "default", r=r, lora_alpha=lora_alpha, lora_dropout=lora_dropout)
-    module.k_proj = peft_cls(module.k_proj, "default", r=r, lora_alpha=lora_alpha, lora_dropout=lora_dropout)
-    module.v_proj = peft_cls(module.v_proj, "default", r=r, lora_alpha=lora_alpha, lora_dropout=lora_dropout)
-    module.o_proj = peft_cls(module.o_proj, "default", r=r, lora_alpha=lora_alpha, lora_dropout=lora_dropout)
-    return module
+@pytest.fixture()
+def attention_inputs(seed: int=42, device: torch.device='cuda'):
+    torch.manual_seed(seed)
 
-def load_gptq_model_weights(model_name, peft_config):
-    from peft.tuners import LoraModel
-
-    from auto_gptq import (
-        AutoGPTQForCausalLM,
-        BaseQuantizeConfig,
-    )
-    from auto_gptq.utils.peft_utils import (  # pylint: disable=import-outside-toplevel,import-error
-        GPTQLoraModel,
-        get_gptq_peft_model,
-    )
-    from fms_acceleration_peft.autogptq_utils import (  # pylint: disable=import-outside-toplevel
-        create_new_module_peft,
-        replace_module_peft,
-    )
-    quantize_config = BaseQuantizeConfig.from_pretrained(model_name)
-    model = AutoGPTQForCausalLM.from_quantized(
-                model_name,
-                quantize_config=quantize_config,
-                use_tritonv2=True,
-                trainable=True,
+    inputs = {}
+    for dtype in DTYPES:
+        for bs, seq_len, dim_size in SIZES:
+            inputs[(bs, seq_len, dim_size, dtype)] = (
+                torch.rand(
+                    (bs, seq_len, dim_size), 
+                    dtype=getattr(torch, dtype),
+                    device=device,
+                ),
+                torch.arange(seq_len).repeat(bs,1).to(device)
             )
-    model.model.model.layers = model.model.model.layers[:1]
+    yield inputs
 
-    _old_create_new_module = LoraModel._create_new_module
-    _old_replace_module = GPTQLoraModel._replace_module
-    _create_new_module = partial(create_new_module_peft, target_cls=QuantLinear)
-    LoraModel._create_new_module = staticmethod(_create_new_module)
-    GPTQLoraModel._replace_module = MethodType(replace_module_peft, GPTQLoraModel)
+@pytest.fixture()
+def model_inputs(seed: int=42, device: torch.device='cuda'):
+    torch.manual_seed(seed)
 
-    # Install GPTQ adapters using the AutoGPTQ package (with the above patches)
-    model = get_gptq_peft_model(
-        model,
-        peft_config=peft_config,
-        auto_find_all_linears=peft_config.target_modules is None,
-        train_mode=True,  # install adapaters for training
-    )
-
-    LoraModel._create_new_module = staticmethod(_old_create_new_module)
-    GPTQLoraModel._replace_module = MethodType(_old_replace_module, GPTQLoraModel)
-
-    return model.model.model.layers[0].self_attn
-
-def build_model(model_name, base_type, r, lora_alpha, lora_dropout, dummy=True):
-    def get_general_kwargs(mod):
-        return {
-            "input_features": mod.in_features,
-            "output_features": mod.out_features,
-            "bias": mod.bias is not None,
-        }
-    if base_type == BNB:
-        config = AutoConfig.from_pretrained(model_name)
-        quant_cls = Linear4bit
-        peft_cls = LoraBNBLinear4bit
-        base_type_kwargs = {
-            "compute_dtype": torch.float16,
-            "quant_type": 'nf4',
-            "quant_storage": torch.float16,
-        }
-    elif base_type == GPTQ:
-        config = AutoConfig.from_pretrained(model_name)
-        quant_cls = QuantLinear
-        peft_cls = LoraGPTQLinear4bit
-        base_type_kwargs = {
-            "bits": 4,
-            "group_size": -1,
-        }
-    else:
-        raise NotImplementedError
-
-    if dummy:
-        test_module = LlamaAttention(config, 0)
-        test_module.q_proj = quant_cls(
-                        **get_general_kwargs(test_module.q_proj),
-                        **base_type_kwargs
-                    )
-        test_module.k_proj = quant_cls(
-                        **get_general_kwargs(test_module.k_proj),
-                        **base_type_kwargs
-                    )
-        test_module.v_proj = quant_cls(
-                        **get_general_kwargs(test_module.v_proj),
-                        **base_type_kwargs
-                    )
-        test_module.o_proj = quant_cls(
-                        **get_general_kwargs(test_module.o_proj),
-                        **base_type_kwargs
-                    )
-        return wrap_peft(test_module, peft_cls, r, lora_alpha, lora_dropout)
-    else:
-        peft_config = LoraConfig(
-            r=r, 
-            lora_alpha=lora_alpha, 
-            lora_dropout=lora_dropout, 
-            target_modules=["q_proj","k_proj","v_proj","o_proj"]
-        )
-        if base_type == BNB:
-            quant_config = BitsAndBytesConfig(
-                load_in_4bit=True,
-                bnb_4bit_quant_type=base_type_kwargs["quant_type"],
-                bnb_4bit_compute_dtype=base_type_kwargs["compute_dtype"],
-                bnb_4bit_use_double_quant=True,
+    inputs = {}
+    for dtype in DTYPES:
+        for bs, seq_len, dim_size in SIZES:
+            inputs[(bs, seq_len, dim_size, dtype)] = (
+                torch.randint(
+                    0, 10000, # most models should have more than 10K
+                    (bs, seq_len), dtype=torch.int, device=device,
+                ),
+                torch.randint(
+                    0, 10000, # most models should have more than 10K
+                    (bs, seq_len), dtype=torch.int, device=device,
+                ),
+                torch.arange(seq_len).repeat(bs,1).to(device)
             )
-            config.num_hidden_layers = 1
-            model = AutoModelForCausalLM.from_pretrained(
-                model_name,
-                config = config,
-                torch_dtype = torch.float16,
-                quantization_config=quant_config,
-            )
-            model = get_peft_model(model, peft_config)
-            return model.model.model.layers[0].self_attn
-        elif base_type == GPTQ:
-            return load_gptq_model_weights(model_name, peft_config)
-        else:
-            raise NotImplementedError("Invalid base type")
+    yield inputs
 
-def prepare_attn_inputs(bs, seq_len, dim_size):
-    x =  torch.rand((bs, seq_len, dim_size), dtype=torch.float16, device="cuda")
-    position_ids = torch.arange(seq_len, device=x.device).view(1, -1)
-    return x, position_ids
+@pytest.fixture()
+def dropout_masks(seed: int=42, device: torch.device='cuda'):
+    torch.manual_seed(seed)
 
-def prepare_model(base_peft_model, base_type, dropout_mask, foak_patch=False):
-    _model = base_peft_model.to("cuda")
+    masks = {}
+    for d in DROPOUTS:
+        binomial = torch.distributions.binomial.Binomial(probs=1-d)            
+        for _, sl, hid in SIZES:
+            if (sl, hid) not in masks:
+                masks[(sl, hid, d)] = binomial.sample((sl, hid)).to(device)
 
-    # Inject Dummy Dropout Here
-    _model.q_proj.lora_dropout = torch.nn.ModuleDict([["default", DummyDropout(dropout_mask)]])
-    _model.k_proj.lora_dropout = torch.nn.ModuleDict([["default", DummyDropout(dropout_mask)]])
-    _model.v_proj.lora_dropout = torch.nn.ModuleDict([["default", DummyDropout(dropout_mask)]])
-    _model.o_proj.lora_dropout = torch.nn.ModuleDict([["default", DummyDropout(dropout_mask)]])
+    yield masks
 
-    if foak_patch:
-        if base_type == BNB:
-            # TODO sets quant_state dtype to float16 here 1st to avoid type mismatch error
-            # will revisit later to find out how to initialize quant_state on casting to device
-            base_peft_model.q_proj.base_layer.weight.quant_state.dtype = torch.float16
-            base_peft_model.k_proj.base_layer.weight.quant_state.dtype = torch.float16
-            base_peft_model.v_proj.base_layer.weight.quant_state.dtype = torch.float16
-            base_peft_model.o_proj.base_layer.weight.quant_state.dtype = torch.float16
+@pytest.fixture()
+def attention_layers(device: torch.device = 'cuda'):
+    from bitsandbytes.nn.modules import Linear4bit
+    # from auto_gptq.nn_modules.qlinear.qlinear_tritonv2 import QuantLinear
 
-        _model = patch_model(base_peft_model, base_type=base_type)
-        print(patch_model_summary())
+    QUANTIZED_BASE_LAYER_CLASSES = {
+        BNB: Linear4bit
+    }
 
+    # this is only done for BNB
+    layers = {}
+    base_type = BNB
+    for dtype in DTYPES:
+        for r, lora_alpha in LORA_PARAMS:
+            model_name, attn_cls, target_modules = TEST_MODELS[base_type]
+            quant_cls = QUANTIZED_BASE_LAYER_CLASSES[base_type]
+            peft_cls = LORA_QUANTIZED_CLASSES[base_type]
+
+            if base_type == BNB:
+                base_type_kwargs = {
+                    "compute_dtype": getattr(torch, dtype),
+                    "quant_type": 'nf4', # NOTE: fp4 is not supported by atm
+                    "quant_storage": getattr(torch, dtype),
+                }
+            elif base_type == GPTQ:
+                base_type_kwargs = {"bits": 4,"group_size": -1}
+            
+            # use the llama model
+            config = AutoConfig.from_pretrained(model_name)
+            attn_module = attn_cls(config, layer_idx=0)
+            for tm in target_modules:
+                mod = getattr(attn_module, tm)
+                quant_base_layer = quant_cls(
+                    input_features=mod.in_features,
+                    output_features=mod.out_features,
+                    bias=mod.bias is not None,
+                    **base_type_kwargs,
+                )
+
+                if base_type == BNB:
+                    # need to cast it so that the quant_type will be correct
+                    # because it takes the quantype from the weights dtype
+                    quant_base_layer.to(getattr(torch, dtype))
+
+                # bring to device
+                quant_base_layer = quant_base_layer.to(device)
+
+                lora_linear_layer = peft_cls(
+                    quant_base_layer, ADAPTER_NAME, 
+                    r=r, lora_alpha=lora_alpha, 
+                    lora_dropout=0. # will override the dropout anyway
+                )
+
+                # this means all target modules get the same dropout
+                lora_linear_layer.lora_dropout = torch.nn.ModuleDict(
+                    [[ADAPTER_NAME, DummyDropout()]]
+                )
+                setattr(attn_module, tm, lora_linear_layer)
+
+            layers[(base_type, r, lora_alpha, dtype)] = attn_module.to(device)
+
+    yield layers
+
+@pytest.fixture()
+def loaded_models(device: torch.device = 'cuda'):
+
+    from fms_acceleration_peft.framework_plugin_bnb import BNBAccelerationPlugin
+    from fms_acceleration_peft.framework_plugin_autogptq import AutoGPTQAccelerationPlugin
+
+    plugins = {
+        BNB: BNBAccelerationPlugin(
+            {
+                'peft': {
+                    'quantization': {
+                        'bitsandbytes': {'quant_type': 'nf4', 'no_peft_model': False}
+                    }
+                }
+            }
+        ),
+        GPTQ: AutoGPTQAccelerationPlugin(
+            {
+                'peft': {
+                    'quantization': {
+                        'auto_gptq': {'kernel': 'triton_v2', 'from_quantized': True}
+                    }
+                }
+            }
+        ),
+    }
+
+    class TrainArgs:
+        gradient_checkpointing = False
+        gradient_checkpointing_kwargs = {}
+    args = TrainArgs()
+
+    all_models = {}
+    for dtype in DTYPES:
+        for base_type in [BNB, GPTQ]:
+            for r, lora_alpha in LORA_PARAMS:
+                model_name, _, target_modules = TEST_MODELS[base_type]
+                # config = AutoConfig.from_pretrained(model_name)
+                peft_config = LoraConfig(
+                    r=r, 
+                    lora_alpha=lora_alpha, 
+                    lora_dropout=0., # anyway we are going to override it
+                    target_modules=target_modules,
+                )
+
+                _plugin = plugins[base_type]
+                model = _plugin.model_loader(
+                    model_name, torch_dtype=getattr(torch, dtype)
+                )
+                model, _ = _plugin.augmentation(model, args, (peft_config,))
+                all_models[(base_type, r, lora_alpha, dtype)] = model
+
+                lora_mods = []
+                for mod in model.modules():
+                    if hasattr(mod, 'lora_dropout'):
+                        lora_mods.append(mod)
+                
+                for mod in lora_mods:
+                    mod.lora_dropout = torch.nn.ModuleDict(
+                        [[ADAPTER_NAME, DummyDropout()]]
+                    )
+
+    return all_models
+
+def prepare_foak(attn_module, base_type):
+    _model = patch_model(attn_module, base_type=base_type)
+    # print(patch_model_summary())
     return _model
 
-def run_model(model, X, position_ids):
-    with torch.autocast(dtype=torch.float16, device_type="cuda"):
-        out, _, _ = model(X, position_ids=position_ids)
+def run_model(
+    model, dtype, X, 
+    device: torch.device = 'cuda', **kwargs, 
+):
+    with torch.autocast(dtype=getattr(torch, dtype), device_type=device):
+        outputs = model(X, **kwargs)
+
+    if hasattr(outputs, 'loss'):
+        out = outputs.loss
+    else:
+        out = outputs[0]
     loss = out.norm()
     loss.backward()
-    grad_norms_A = [
-        model.q_proj.lora_B.default.weight.grad,
-        model.k_proj.lora_B.default.weight.grad,
-        model.v_proj.lora_B.default.weight.grad,
-        model.o_proj.lora_B.default.weight.grad,
-    ]
-    grad_norms_B = [
-        model.q_proj.lora_B.default.weight.grad,
-        model.k_proj.lora_B.default.weight.grad,
-        model.v_proj.lora_B.default.weight.grad,
-        model.o_proj.lora_B.default.weight.grad,
-    ]
 
-    return loss, (torch.vstack(grad_norms_A), torch.vstack(grad_norms_B))
+    return loss
 
-TEST_MODELS = {
-    BNB: "TinyLlama/TinyLlama-1.1B-Chat-v0.3",
-    GPTQ: "TheBloke/TinyLlama-1.1B-Chat-v0.3-GPTQ",
-}
+def get_modules_with_class(model, cls):
+    modules = []
+    name = cls.__name__
+    for mod in model.modules():
+        # check for the class name in the MRO
+        mro = mod.__class__.mro()
+        if any(name == x.__name__ for x in mro):
+            modules.append(mod)
+
+    if len(modules) == 0:
+        raise ValueError(f"cannot find modules with class '{cls}'")
+    return modules
+
+def get_attention_lora_grads(
+    model, target_modules
+):
+    # comparing the grads on the adapters
+    adapter_grads = []
+    for tm in target_modules:
+        for n in ['lora_A', 'lora_B']:
+            mod = model.get_submodule(f"{tm}.{n}.{ADAPTER_NAME}")
+            adapter_grads.append(mod.weight.grad) 
+    if len(adapter_grads) == 0:
+        raise ValueError(f"cannot find adapter grads")
+    return adapter_grads
+
+# -------------------------- TESTS ----------------------------------
 
 # small
-def test_adapter_gradients_match_with_dummy_module():
+def test_adapter_gradients_match_with_attention_layer(
+    attention_inputs, attention_layers, dropout_masks
+):
     '''
-    this function loads a dummy attention module, 
-    For GPTQ it isn't straightforward to initialize a dummy baselayer, 
-    so this function is only supports BNB for now
+    Construct and test equivalence on a single constructed attention
+    module.
+    - For GPTQ it seems to be troublesome to initialize an insolated
+      layer, thus this test is only done for BNB.
     '''
-    torch.manual_seed(42)
-    rtol = 1e-3
-    X, position_ids = prepare_attn_inputs(bs=1, seq_len=128, dim_size=2048)
-
     for base_type in [BNB]:
-        print(base_type)
-        for adapter_dropout in [0, .1, .5]:
-            # Prepare fixed test inputs
-            # build base model
-            base_peft_model = build_model(
-                TEST_MODELS[base_type], base_type, r=8, lora_alpha=1, lora_dropout=adapter_dropout, dummy=True,
-                )
-            binomial = torch.distributions.binomial.Binomial(probs=1-adapter_dropout)            
-            dropout_mask = binomial.sample(X.shape[1:]).to(X.device) * (1.0/(1-adapter_dropout))
+        model_name, _, target_modules = TEST_MODELS[base_type]
+        config = AutoConfig.from_pretrained(model_name)
 
-            # prepare 2 model variants to compare before and after foak patching
-            unpatched = prepare_model(deepcopy(base_peft_model), base_type, dropout_mask, foak_patch=False)
-            patched = prepare_model(deepcopy(base_peft_model), base_type, dropout_mask, foak_patch=True)
-            # run the models and get the loss and gradients
-            loss_unpatched, grads_unpatched = run_model(unpatched, X, position_ids)
-            loss_patched, grads_patched = run_model(patched, X, position_ids)
-            assert (loss_unpatched - loss_patched).abs() < 1e-5,\
-                "Loss after foak patch do not match"
+        # find compatible sizes
+        sizes = [(bs, sl, hd) for bs, sl, hd in SIZES if hd == config.hidden_size]
 
-            A1, B1 = grads_unpatched
-            A2, B2 = grads_patched
+        for (bs, sl, hd), dtype in product(sizes, DTYPES):
+            X, position_ids = attention_inputs[(bs, sl, hd, dtype)]
+            _kwargs = {'position_ids': position_ids}
 
-            for b1, b2 in zip(A1 + B1, A2 + B2):
-                assert torch.allclose(b1, b2, atol=1e-4, rtol=rtol), "Gradients don't match after foak patch"
+            for r, lora_alpha in LORA_PARAMS:
+                for d in DROPOUTS:
 
-# large
-def test_adapter_gradients_match_with_model_weights():
-    '''
-    this function initializes a full model with actual model weights
-    and runs the attention module from the 1st layer
-    '''
+                    # attn layer + mask
+                    attn = attention_layers[(base_type, r, lora_alpha, dtype)]
+                    DummyDropout.dropout_mask = dropout_masks[(sl, hd, d)]
+
+                    # prepare models
+                    without_foak = deepcopy(attn)
+                    with_foak = prepare_foak(deepcopy(attn), base_type)
+
+                    # run the models and get the loss and gradients
+                    loss_unpatched = run_model(
+                        without_foak, dtype, X, **_kwargs
+                    )
+                    loss_patched = run_model(
+                        with_foak, dtype, X, **_kwargs
+                    )
+
+                    # compute outputs
+                    assert (loss_unpatched - loss_patched).abs() < LOSS_TOL,\
+                        "Loss after foak patch do not match"
+
+                    grads_unpatched = get_attention_lora_grads(without_foak, target_modules)
+                    grads_patched = get_attention_lora_grads(with_foak, target_modules)
+
+                    for x, y in zip(grads_unpatched, grads_patched):
+                        assert torch.allclose(x, y, atol=ALLCLOSE_ATOL, rtol=ALLCLOSE_RTOL), "Gradients don't match after foak patch"
+
+@pytest.mark.slow
+def test_adapter_gradients_match_with_model(
+    model_inputs, loaded_models, dropout_masks
+):
+    """
+    Using full models loaded by plugins, test equivalence on random inputs.
+    """
+
     for base_type in [BNB, GPTQ]:
-        print(base_type)
-        for adapter_dropout in [0.1, 0.5]:
-            # Prepare fixed test inputs
-            X, position_ids = prepare_attn_inputs(bs=1, seq_len=128, dim_size=2048)
-            base_peft_model = build_model(
-                TEST_MODELS[base_type], base_type, r=8, lora_alpha=1, lora_dropout=adapter_dropout, dummy=False
-                )
-            # prepare 2 model variants to compare before and after foak patching
-            binomial = torch.distributions.binomial.Binomial(probs=1-adapter_dropout)            
-            dropout_mask = binomial.sample((X.size(1), X.size(2))).to(X.device) * (1.0/(1-adapter_dropout))
-            
-            unpatched = prepare_model(deepcopy(base_peft_model), base_type, dropout_mask, foak_patch=False)
-            patched = prepare_model(deepcopy(base_peft_model), base_type, dropout_mask, foak_patch=True)
-            # run the models and get the loss and gradients
-            loss_unpatched, grads_unpatched = run_model(unpatched, X, position_ids)
-            loss_patched, grads_patched = run_model(patched, X, position_ids)
-            assert (round(loss_unpatched.item(),5) == round(loss_patched.item(),5)), "Loss after foak patch don't match"
-            A1, B1 = grads_unpatched
-            A2, B2 = grads_patched
-            # check the gradients on the adapters match with some tolerance before patching and after
-            match_gradients_A = (~torch.isclose(A1, A2, rtol=1e-03, atol=1e-03)).sum() == 0
-            match_gradients_B = (~torch.isclose(B1, B2, rtol=1e-03, atol=1e-03)).sum() == 0
+        model_name, attn_cls, target_modules = TEST_MODELS[base_type]
+        config = AutoConfig.from_pretrained(model_name)
 
-            assert match_gradients_A, "Gradients on lora A do not match after foak patch"
-            assert match_gradients_B, "Gradients on lora B do not match after foak patch"
+        # find compatible sizes
+        sizes = [(bs, sl, hd) for bs, sl, hd in SIZES if hd == config.hidden_size]
+
+        for (bs, sl, hd), dtype in product(sizes, DTYPES):
+            input_ids, labels, position_ids = model_inputs[(bs, sl, hd, dtype)]
+            _kwargs = {
+                'position_ids': position_ids, 'labels': labels
+            }
+
+            for r, lora_alpha in LORA_PARAMS:
+                for d in DROPOUTS:
+
+                    # attn layer + mask
+                    model = loaded_models[(base_type, r, lora_alpha, dtype)]
+                    DummyDropout.dropout_mask = dropout_masks[(sl, hd, d)]
+
+                    # prepare models
+                    without_foak = deepcopy(model)
+                    with_foak = prepare_foak(deepcopy(model), base_type)
+
+                    # run the models and get the loss and gradients
+                    loss_unpatched = run_model(
+                        without_foak, dtype, input_ids, **_kwargs
+                    )
+                    loss_patched = run_model(
+                        with_foak, dtype, input_ids, **_kwargs
+                    )
+
+                    # compute outputs
+                    assert (loss_unpatched - loss_patched).abs() < LOSS_TOL,\
+                        "Loss after foak patch do not match"
+
+                    for _without, _with in zip(
+                        get_modules_with_class(without_foak, attn_cls),
+                        get_modules_with_class(with_foak, attn_cls),
+                    ):
+                        grads_unpatched = get_attention_lora_grads(
+                            _without, target_modules
+                        )
+                        grads_patched = get_attention_lora_grads(
+                            _with, target_modules
+                        )
+
+                    for x, y in zip(grads_unpatched, grads_patched):
+                        assert torch.allclose(x, y, atol=ALLCLOSE_ATOL, rtol=ALLCLOSE_RTOL), "Gradients don't match after foak patch"
